+++ components/accelerator/nvidia/query/clock-events-state/options.go
