+++ components/accelerator/nvidia/query/clock-events-state/state_test.go
